Use numpy's eye and identity tools
np.eye returns a 2-D array with 1's as the diagonal and 0's elsewhere. The diagonal can be main, upper or lower depending on the optional "k =" parameter. A positive k is for the upper diagonal, a negative k is for the lower, and a 0 (default) is for the main diagonal.